Extract draft lifecycle routes out of submissions.ts
Moves the creator self-service routes for a submission's lifecycle (share,
abandon, delete-game) into a new apps/api/src/creation/draft-lifecycle-routes.ts
module. All three share the same shape: verify token, check ownership
against the store, mutate, invalidate caches.

backendFor/builderOf/releaseWorkspace/invalidateStatusCache/
invalidatePublishedGameCaches stay in submissions.ts (shared with many other
call sites) and are passed in as injected options, same pattern as the prior
batches' checkUserAccess/ensureSubmissionSlug.

submissions.ts drops from 4497 to 4323 lines; the new file is 219 lines,
0 comment-prose debt.
